requirements.txt
torch
torchaudio
soundfile
numpy
pydub
ffmpeg-python
loguru
icecream
librosa
noisereduce
onnxruntime
